Adds spree correct class and modules for promotion
