+++ config/initializers/spree.rb
@@ -1,6 +1,6 @@
 Spree::PermittedAttributes.user_attributes.push :referral_code, :affiliate_code
 config = Rails.application.config
 config.after_initialize do
-  config.spree.promotions.rules << Spree::ReferredPromotionRule
-  config.spree.promotions.rules << Spree::AffiliatedPromotionRule
+  config.spree.promotions.rules << Spree::Promotion::Rules::ReferredPromotionRule
+  config.spree.promotions.rules << Spree::Promotion::Rules::AffiliatedPromotionRule
 end
